pyproject.toml
[project]
name = "haystack-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "flask>=3.1.2",
    "haystack-ai>=2.20.0",
    "ollama-haystack>=5.3.0",
    "polars>=1.35.2",
    "sentence-transformers>=5.1.2",
]
